requirements.txt
absl-py==2.0.0
accelerate==0.24.1
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.4.6
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
diffusers==0.21.4
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.7.0
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.3
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.23.4
google-auth-oauthlib==1.0.0
gradio==4.31.0
gradio_client==0.16.2
grpcio==1.59.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface==0.0.1
huggingface-hub==0.23.0
idna==3.4
imageio==2.34.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.5.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.9.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.23.0
opencv-python==4.8.1.78
opencv-python-headless==4.9.0.80
orjson==3.10.3
packaging==23.2
pandas==2.0.3
parso==0.8.3
patchify==0.2.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.7.1
pydantic_core==2.18.2
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==1.7.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.35.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.4
safetensors==0.4.0
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.1.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.5.10
timm==0.9.10
tokenizers==0.14.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.0
torchmetrics==0.11.4
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
triton==2.1.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2023.3
ujson==5.9.0
urllib3==2.0.7
uvicorn==0.29.0
uvloop==0.19.0
wandb==0.17.0
watchfiles==0.21.0
wcwidth==0.2.9
websockets==11.0.3
Werkzeug==3.0.1
xformers==0.0.22.post7
yarl==1.9.2
zipp==3.17.0
